Use absolute times rather than intervals for the SyncTimer durations
In both the jack process call and in SyncTimerThread's time counting
implementations, previously we would use an interval, adding that onto
the previous time. This worked fine with pure integer results, but we
will not always have that luxury. Instead, using absolute time
calculations, we ensure that the period we are waiting for is actually
the correct time. It is vaguely more expensive, but produces
significantly more correct results, and rounding errors have a tendency
to compound over time, and as time is what we're dealing with here,
well, let's try and avoid that.