Use a slightly more correct (and also consistent) method for adding extra ticks
This ensures that we're using the same calculation basis for adding
those extra ticks as we're using for the rest of the code. This is
likely to not matter a great deal functionally, but consistency in
implementation is good, and it also gives us a little bit of extra
statistical information to read out, which is nice (the number of extra
ticks is inherently counted now, so we can spit that out along with the
other information).