Optimise the heck out of zl's jack recording level process call
This moves things like the ui update callback (which calls into python
and further into qml) out of jack's very tightly timed processing
thread. It also changes some of the allocation logic, so that we have
statics where we can, and inline the functions doing the calculation
work (which might well happen on some optimisation levels, but not on
all, so useful to actively ask for that to happen).

End result: No xrun warnings caused by the libzl jack client :)