Add "dispatch immediately" buffer, use it to queue all off notes on stop
This reintroduces SyncTimer's ability to send midi messages. This should
probably be turned back into the thing we use, rather than RtMidi's midi
out (i suspect this will have less latency, though that will need
investimagating)

This also removes the last vestiges of the juce midi output from
SyncTimer (and consequently the requirement on the external tool we
used for the bridge). Cleanup is good.