CAS: add audioLevel, position props, switch to SamplerSynth playback, move juce::Timer to pimpl
Big refactor with other things mixed in, not the prettiest.

The first two are a couple of just properties, one which returns the
audio level (so you don't need to register a callback), and the other
returns the first playback position (also so you don't need a callback).

All playback is now done through SamplerSynth instead of through
tracktion's transport system. This allows us to, in the future, be more
precise about the playback positions, which in turn will allow us to
treat jack as the canonical time reference. For now, though, it should
just not have any visible/audible side effects.

Also throttle the position update to 10 updates per second. This is
crazy low, but also means we get less tearing (though we probably want
to actually check whether we've got compositing enabled and then use 60
updates per second instead).

Finally, moving the juce::Timer work to the pimpl means we don't need
the objects to live in the juce message loop thread. The actual move
will come in a different commit.