SamplerSynthVoice: First playback-resetting system for syncing loops
This is the first step in our "let's make sure loops are looping in sync
with the timer". The next step is to not adjust jack's playhead, but
rather adjust synctimer's timekeeping to match jack. While this will
make the real-time synchronisation weird, it will make the playback more
correct according to jack, which is more important.