ZLRouter: Add configuration reading, and a specific Sampler destination
Configuration is now read based on what zynthian tells us. The
route-to-outputs option (the filter outputs one) seems oddly clashing
with our internals, but if we ship it disabled by default and people
want it, it's not really in the way, so... let's just do that.

We still need to hook up something to make it reload the config when
it changes, only loading on startup is a little silly, but this code
will allow for someone external to do that, and it seems like we
probably want to do that somewhere from the python-via-components side
anyway.

Finally, there's a Sampler destination now, which basically just means
"throw stuff through the passthrough, but nothing else", contrasting
with the no destination option, which turns off all the output for that
channel, including passthrough. This allows us to just... leave
Zynthian's connections in place, so we don't have to mute the midi there
any longer, which took a little while, and was kind of ugly anyway,
conceptually, so, nice.