ZLRouter: Add some more outputs, for other-side filtering purposes
This allows us to listen to the destination without having to listen and
filter ports explicitly on the other end. In short: It allows for
PlayGridManager to listen to the notes that will end up at some specific
destination (or comes from some origin) and react to them specifically,
such as for telling the UI what notes are active on any hardware input,
or currently active on a hardware output, for example.