SyncTimer: Fully treat Jack as the canonical time reference
This is still not quite completed, but we now no longer internally
synchronise jack and synctimerthread, but rather we let synctimerthread
catch up to jack on the regular. This will result in marginal jumpiness
in synctimer, but crispiness in jack, which is where we need it.

We also now have jack send out all of everything, which means both the
sample commands (of which there is now only the one set of queues,
rather than three separate ones).

Also expose jackPlayhead in a getter, with warning not to use it unless
you Know What You Are Doing:tm:, because, well, it's not really supposed
to be the one people use.