ZLRouter: Handle channels changing between note on and off, and signal device add/remove
The latter of the two was basically an easy-to-produce side effect of
splitting all the device inputs to have more information in the form of
a class holding the information.

The former is the important bit, and ensure that once a midi note on
event has been sent to a specific channel from a specific input, then
any subsequent on and off events go the same place, until the note is
entirely off again. It's not even hugely expensive, which is nice.