SamplerSynth(Voice): Switch playback logic from monolithic tracktion to channeled jack
With this, each track (and also two global one, designated effected and
uneffected, for globally played sounds with and without effects added)
has their own jack ports for left and right, which means we are able to
pipe it through individual effects stacks. For now everything just plays
to the system playback ports, but this is the required groundworks for
making that possible.

It also, for now, switches the voice logic to 16 per track instead of
128 global ones. This might need a little extra work, but also 16 is
kind of a lot for a single track. Kind of. Maybe it needs to be 12 per
sample or something like that. We'll have to see as we go along, it's
easy to adjust.