SamplerSynth: Forward jack timing to voice, disable sync for now
The previous implementation of sample playback synchronisation wouldn't
work for jack (and was originally intended for just synchronising
between jack and tracktion). Playback works better for now without it,
so seems sensible to just get rid of that old one for now, while i work
out how to implement a proper one.