SamplerSynthVoice: Rework the beat-matching looping logic
Each time we run, we bind ourselves to the timer ticks, and the local
length of a subbeat in microseconds, so we'll never be further wrong
than a single loop before we end up correct again. This might actually
be kind of interesting for effects purposes at some point, but right now
it just means that the looping is sane.