CAS/SamplerSynth: Read volume directly out of CAS, otherwise don't use
This means that loops are now set to full volume, but SamplerSynthVoice
now honours the volume set on a ClipAudioSource object, so we
effectively have two volume sets - one for the clip itself, and one
per-note. Seems like what we actually want anyway, so, slightly simpler
to understand code, and more power. Nice when that happens.