SamplerSynth: Optimize (less voices, clear buffer w memset, pointering)
Storing the samplersynth tracks as pointers now. Initially intended to
use the ability to store data in the tracks per process call, but ended
up not doing that as the particular thing didn't make sense.

Now using memset to clear the audio buffers before sending them into the
voices. Potentially we could do some magic here to not require clearing
the buffers each run if there was nothing done last run, but also jack
says that you /should/ clear them if you want silence out, so not sure
we can really get away with that, and we'll just have to deal with the
perpetual cost.

We also now have less voices per track. Perhaps we could do some magic
and use a pool for starting and stopping, but that would result in list
manipulation during playback, which we would probably want to not do, so
while it would be more flexible, it would also be a problem.