SyncTimer: If operating on an uncleared buffer, insert at the end
To ensure we don't attempt to insert events at a time before the most
recently inserted event, let's ensure we set the first available frame
to the next frame after the last event in the buffer, if we are
operating on a buffer that MidiRouter hasn't cleared for us. This is a
fairly common occurance, and i suspect there are other places where the
"output outputs, input reads and then clears" logic is going to need to
be applied to ensure we never drop notes. However, i am also unsure if
it's a guaranteed situation that we have here, or whether it's because
it's a direct connection, we have the same buffer object at both ends
(like, what happens when there's multiple connections to a single input
port? seems like it could be a thing to consider)