SyncTimer: timerThread var to global, and experimentally remove step's isPlaying
The former has a similar effect to what's found in MidiRouter, and i
similarly am not sure why. The other change, removing the isPlaying
logic from the ring's steps, is partially that i'm pretty sure we just
don't need it (it would warn before, and i never did see that warning
pop up), but also it seemed to come up as a cause of __dabt_usr
situations.