AudioLevels: Remove a few unneeded bits
This also attempts to reduce the number of __dabt_usr situations,
however for some reason which entirely escapes me, the two first lines
in the process call (that is, the two which get the port buffers and
assign them to a local variable) show up in hotspot as having vastly
different costs (second being about ten times more time), which
strongly suggests that this is hitting __dabt_usr, and i do not see how
and where and so on. Very frustrating.