MidiRouter: Add explicit port for zynthian and external, and rejig debug for write event
The latter is just a touch of reworking to make sure the function flows
more sensibly (and also hits the debug for retries)

The former is the meat, and means we no longer have a midi mixdown
situation further out. Since jack's code itself suggests that midi
mixdown is very expensive, it's worth avoiding it if we can, and since
we can, let's do that.

Leaving some old stuff commented out for now, but can probably get rid
of that.