AudioLevels: An attempt to work around an occasional broken-port situation
This does not fix the problem, rather plasters over it a bit. It will
still cause issues, but we no longer spam that "oh no invalid port"
message all the time, which is very expensive... Will need to look at
this again at some point (hence the debug output on disabling and the
incomplete fixes included here) but... yeah, this wants more attention
at some point.