AudioLevels: Rework some of the internals (singleton, timer, init, debug)
No longer attempt to replace the jack client when it fails to initialise
properly and we end up with incorrect ports. This is a partial solution
and the issue (broken client) will need fixing properly, but the
replace-client workaround was the wrong fix anyway, so ripping that out
with the intention to fix it properly instead.

Better debug output for the initialisation steps, because the previous
output was kind of odd and confusing, and it now is more like what the
rest of the system does.

Switch to a fully thread safe singleton method. Not our problem, it
turns out, but also it's not bad to have, so leaving that in.

We switch to using a qtimer instead of a juce timer. Functionally the
same, but it means we only have one superclass.