SyncTimer: Keep a midibuffer instance around, avoid a new in process()
We're getting closer to no memory stuff in SyncTimer's process call. Now
that we've established that the idle time weirdness was not in fact from
here, we can begin to work more on improving the memory management
aspects of the process call itself. There's more to do, and less low
hanging fruit, but also it'll be very good indeed to get these things
sorted.