SyncTimer/Commands: Reuse old ones instead of delete/new, increase pool size
Suspect we might need to increase the pool size further for particularly
bonkers tracks, but this should do for now at least. We've got plenty of
memory space to grow anyway, the objects are quite small, so that's not
much of a problem, but no reason to make a huge pool until it's a
problem we actually have.

In the meantime, what this change actually does is mean no memory
allocation work during playback, which is nice. Only thing that happens
is the deletion queue still being a QList with a varying size, but we
can deal with that if it turns out to actually be a problem.