SyncTimer: Adjust step handling to 96ppqn, and a bit of detail adjustment
The way we use the timer in places now means that we should, in fact, be
able to adjust this basically as high as we want it, but also processing
cost is of course a thing and for now, 96ppqn seems reasonable given
everything else.

However, the most important thing here is that it gives us a pile of
flexibility we didn't before: At 32ppqn, the potential divisions are
quite narrow (1, 2, 4, 8, 16, and 32), whereas for 96 this list is 1, 2,
3, 4, 6, 8, 12, 16, 24, 32, 48, and 96. The most important one here is
24, which is the speed at which the midi beat sync clock should run at,
and which is conspicuously absent from the list of factors for 32ppqn.
So, in short, we need this for midi sync to work, but also it's just
nicer in general. Wins all 'round ;)