SyncTimer: Calculate quantized BPM for jack transport, also a bit of pretty
The first bit is to ensure that the BPM we inform about as sent out
through the jack transport manager is correct. This is done by doing BPM
changes per-step, and calculating what the BPM is for each step section
inside a period. This seems like kind of a lot of work, but it is
important to ensure that jack's transport is correctly synchronised with
what synctimer is doing. Also, it isn't really /that/ much work, just a
bit of back and forth, so it's not too bad in reality.

Decided to do the TimerCommand operation handling as a switch instead of
a chain of if statements, because well, it had outgrown making sense as
an if chain, so... yup, it was time.