Bindings : Disable return-value-heuristic when generating bindings with shiboken
When return value heuristic is enabled, it is causing the returned objects to be repaarented to the object which it is called from. The reparenting is causing some c++ object to get deleted as the reference were not kept.

In our case we already have well defined parents. If some object does not have a defined parent, it should have a parent and needs fixing.
Enabling this return value heuristic option does not make sense for our case.