SamplerSynth: Add Hermite interp, fix off-by-one, split low/highpass
The low and highpass filters are now applied separately for two given
CCs. By default we're using 1 (mod wheel) for lowpass and 74 (mpe slide)
for highpass, but it's easy to change, and we could also expose both of
those for configuration if we want to.

Added Hermite interpolation for pitch shifting, replacing the old linear
interpolation. This should result in much smoother interpolation when
shifting more than just a couple of semitones. When not pitch shifting,
we just fetch the sample data as given. Hermite interpolation is
reasonably inexpensive, but also if we don't need to do it, it's worth
avoiding (since playing an un-pitch-shifted sample is a common use case,
both for the metronome, and for loop playback, where shifting is done
ahead of time using juce's built-in high quality non-realtime pitch
shifter).

Finally, the main fix here is the off-by-one fix where we used to
increment the output buffer before assigning the sample value (which
effectively means we would end up zeroing one sample per period, which
ended up causing some nasty glitchiness in the output). Easy fix, took a
while to spot such a basic error, but now it's done ;)