MidiRouter: (dis)connect zynthian ports from/to outputs, only create 10
The first one means we are only connecting output ports to zynthian when
the channel is supposed to be playing synth sounds. This fixes the
problem of us playing synthy sounds if the channel was set to being
sample based (or indeed anything other than zynthian, which is the only
time it should be making synth noises).

We also only create 10 output channels now, as Sketchpad only has ten
anyway and there's no reason for those other six channels to exist when
they literally can't be used anyway. Just wasted cpu cycles at that
point and we only have so many of those.