SamplerSynthVoice: Don't try to fetch data outside the sample range
How this doesn't crash is frankly a touch worrying (surely this is
memory we're not supposed to have read access to? But ok). It's already
handled for pitch bent notes, but for the unbent stuff, i missed this
one. Also handle single channel samples (like the metronome clicks), so
they're just panned center instead of only the left channel.