SamplerSynth: Add granular playback option for SamplerSynth
This thing eats voices like there's no tomorrow, but it's also quite
neat. Still some work to be done to really get it tight (the aftertouch
and pitch shifting stuff is not done at the correct position in time,
but that's more a bit of work to do rather than particularly difficult).

As it stands, Granulator will perform playback using a clip's granular
settings, and uses the clip's standard adsr envelope for initial volume
on the grains.

This is going to need a bit of work yet, but it's coming together very
nicely.