SyncTimer: Refactor bpm handling to always be TimerCommands
We no longer set bpm explicitly at the moment the function is called,
rather we set it scheduled by the timer, and then update as we reach
that position in the schedule. This ensures precision (at the cost of a
minute amount of lag), and requires some changes in the consumers as
well, so make sure to update both packages at the same time.

Also make start() just start, don't bother setting the bpm because it kind
of... is just set anyway. This was basically a leftover from way back,
when we didn't have the ability to change bpm on the go and also just
had a basic timer. Things are a bit different now and this was always
given the current bpm anyway, and consequently unnecessary anyway.