SamplerSynth: No more juce::SamplerSynth, adapt to new always-going logic
This is a major bit of reworking, and we now do all our own sample
playback work (we mostly did already, we now also handle the start/stop
ourselfes). This ensure we start sample playback at a frame precision
level, instead of a buffer precision level. As a result, we can now do
granular playback down to the individual frame positioning, and we also
have much tighter control over loop start/stop as a side effect. There
is likely much more to polish on this, but it works for now.