SegmentHandler: StartOffset property, don't update segments except on playback start
This adds in a startOffset property, which gives the value that was
passed to the most recent startPlayback call.

Further, we only update the segments data on playback start now, and
also only for the amount of time the playback will actually run. This
ensures that when we just play one segment, we don't end up playing the
first few ticks of the next segment as well (as that segment never gets
turned on), which makes for a more pleasant experience.