SamplerSynth: Add lane concept, and a global voice pool
We now have 384 global voices, instead of the fixed 32 voices per
channel we used to have (it's the same number of voices, but more
flexible). We likely will need to make this a smaller number to not
completely blow up a raspberry pi 4, but leaving it huge for now to be
able to experiment. If it turns out not to be a problem, then it's not a
problem. If it's a problem, we can change it.