Initial work towards porting the recorder functionality to sketchpad tracks
The concept here is to stop using midi channels as meaning sketchpad
tracks. As we need to be able to handle recording midi events for
multiple tracks simultaneously to make the song recorder work properly
for producing sketches (that is, we want any recorded audio we produce
to also contain the midi events that produced it), this is the most
sensible place to start that work.

We will eventually need to entirely split these things up, so we can
also record mpe events (which will mean a bit of reworking of midirouter
and synctimer, so there's matching ports for each track, but this is
less difficult and more a case of it being just needing some time to
make it happen, and we've got more important things to get done first).