MidiRouter: Don't be clever with sysex
The old logic assumed that sysex would be delivered per-byte instead of
a lump message. Apparently that's just not how jack does it. That,
sort-of-obviously, will break with large sysex messages, but that also
is not really supposed to be how it's done, apparently, so... yeah,
let's maybe just not do that, and then work out what to do with
over-large messages at some later point if that becomes a thing we have
to deal with.