Port MidiRouter to the new MidiRouterDevice class for all i/o purposes
This equalises the way in which messages are handles for input and
output, and greatly reduces the complexity of the implementation. There
should be no performance impact, but it would be advantageous to
investigate this for potential optimisation paths (that is to say, spend
a bit of time with hotspot, see what comes up).