Add individual output ports per track in SyncTimer, change scheduling functions to match
This is a fairly invasive change, and will have partner changes in other
places. The basic concept here is to ensure we can schedule mpe-style
events into a track, and that means we need an output port per-track
instead of a single one (which only allowed one channel per track,
functionally).

Also adds a way of getting a free midi channel for a track, as matches
the schedule and whatnot. This allows us, again, to perform scheduling
in an mpe manner.