MidiRouter: Add support for sub-devices with identical names
Some devices (like the Keystation 61 MK3) connect multiple devices with
identical names, but with different purposes. To ensure we actually
support that, detect those devices as different types of device. This
also adds in the ability to have devices which only registers one input
or output for the same physical device, except different with different
names (such as the korg nanoKey 2, which registers "keyboard" for input,
and "ctrl" for output).