MidiRouterDevice: Add ability to limit output to specific midi channels
This is important for us, as some of the Zynthian synths (in particular
ZynAddSubFX) only support listening on a single specific channel. So,
while this does support multiple channels, it works best with either all
channels or a single channel (we can add the light-squeeze later if it
turns out to be something we actually want).