AudioLevelsChannel: Add support for handling TimerCommands at a timestamp
This allows us to schedule specifically the start of recording into the
AudioLevelsChannel itself, which allows us to first complete one
recording, and then start another recording at basically the same time.

It is entirely possible we'll need to do some work to move some logic
around in DiskWriter, at this point i feel like it's going to be doing
the initialisation work in the processing thread, which we really want
to not be doing if we can at all help it.