AudioLevels: Add the elaborate start/stop recording TimerCommand handling
The original logic required you to do all the setup ahead of time, but
as the new bouncing logic requires us to be able to stop and start
recording at playback time, very rapidly after each other, and with new
filename settings as it goes along, this logic becomes required. See
previous commit on AudioLevelsChannel on some future work that likely
will want to be done sooner rather than later.