Note: Replace setIsPlaying logic with a ref-counted system
The old setIsPlaying logic tried to combine a few too many things, and
did not do much in the way of tracking internally. We now have a set of
three functions for note event registrations, which are used by the
PlayGridManager's event handling to update the Note instances' states to
match reality, which is now reflected more correctly.

This is considerably more solid, and resolves the dual-purpose issue
we've got for the class, by also introducing a separate variable which
tracks the channel fetched for setOn, and uses that instead of the active
channel when handling our internal on/off thing.