Refactor the process load logic, move into MidiRouter, make a property
This allows us to read the information out in QML as well, which is a
handy sort of thing that we kind of want to be able to do anyway
(instead of just depending on the cli output during playback). It also
puts it in a much more reasonable place, instead of the magic nonsense
of having it sitting in the SamplerSynth clients, which was always
supposed to be a stop-gap, and... well, this is much better.