SyncTimer: When asking to send immediately, actually send on next play-step
Prior to this, during playback, the scheduling would happen onto the
next step in the cumulative playback sequence, rather than the next step
to actually be processed for playback. This would cause a
double-the-processing-length delay for e.g. triggering notes on a notes
grid while playback was ongoing. This is much, much tighter. It also
changes the logic a touch for when the missing buffers get submitted (so
they're now next-possible-moment rather than postponed, which ensures
they actually are going to happen in order as expected).