SequenceModel: Change the pattern file naming logic, and improve task messaging
This will also delete the folder for a sequence model if there are no
entries in there.

To restore old sequences, rename the pattern files to a
part(tracknumber)(partletter).pattern.json and then rename the folder
from T1 to global (we won't have any which have the others, but the
logic there becomes having a folder named global2, global3 and so on to
match the scene numbers).