Adapt to new naming logic for playback start/stop
This in fact removes the need for any such thing to happen again in the
future (since we're now just outright telling all sequencemodels to
either start or stop, rather than asking for each one by name). This
should also be considerably less costly, though in reality it doesn't
exactly happen during playback, so to speak, but less strings is more
better, so why not.