PlayfieldManager: A new singleton class which handles the playfield state
This class is designed to handle the playfield state, in a way very
similar to how SegmentHandler currently does for song mode playback. It
is deliberately designed to match that logic, but with a bit more
flexibilty for live playback purposes (so you can set the next state
indead of just the current), which allows us to more gracefully handle
switching between clips in a live setting, where we optimally want to be
doing the state setting in a next-bar type of manner.