bindings: Add PlayfieldManager, plus a couple of missing enums
This allows us to use PlayfieldManager from python, and the enums from
MidiRouter and MidiRecorder that were previously missing are just kind
of nice to have in there, why not. Probably other places where we could
do that, too, just came across these while working on PlayfieldManager,
so thought why not.