SyncTimer: Add function to get the timer tick for a given jack playhead value
This is designed to be used for converting a jack frame time (that is,
the explicit since-jack-started frame position of a given event) into an
equivalent SyncTimer timer tick. This allows us to take any event on the
system, and get a timer position out of it, which immediately is useful
for doing live midi event recording, but which seems to potentially also
be useful for a great many other things.