Add per-pattern swing, per-subnote ratchet and probability
Swing is a full-pattern complement to the microtiming afforded by the
existing per-subnote position parameter. Technically it would allow you
to adjust the positioning of a step's notes all the way past the second
step in the future, but... well, don't do that, but you can, but why
would you? Anyway, swing is a thing now.

Also, per-subnote probability, in the form of a percentage per subnote
that defines how likely it is that that subnote will fire when the
sequencer reaches that step.

Finally, we have a highly granular-optioned ratchet function, in which
you can set three options: The style, how many retriggers you want, and
the probability of each of the retriggers actually being sent for
playback. The probability works like the per-subnote probability, just
per retrigger. The style setting defines whether you calculate the
interval between regriggers using the step size (default) or the
manually set duration, and whether the individual retriggers choke each
other or whether (default) each has a the duration defined for that
subnote.

The probability options require a full rebuild of that step's data on
each run, so the position is invalidated if either of the probability
settings are set.