SyncTimer: Send out all messages on stop, adjust on velocity down hard
This ensure that, as we add in more automation and the like, we'll not
end up causing unnecessary havoc by just ripping out all the stuff that
might actually need to happen in a known order and such. It also means
that while we don't have correct velocity on the on events, we /will/
have fully matching pairs for everything, which is just... very nice.