WaveFormItem: Add support for pointing at an AudioLevelsChannel, and clean up
We can now feed a uri to the source property which instead of a file on
disk will cause the item to show the live-updated thumbnail created by
our DiskWriters while recording. The code is as yet untested, but we
need it in to actually be able to do that testing, and it doesn't impact
the old functionality.

The clean-up includes removing the use of a new thumbnail cache and
format instance for every new thumbnail, which is both wasteful, and
also partially defeats the purpose of the cache itself.

It also includes fixing a variety of small interaction issues, which
would cause updates where there shouldn't be any, and a lack of updates
where there should. Basically, general polish.

One todo still to be considered is, what do we do for multi-channel
sounds? Do we use the "stack them up" style that everybody else uses,
or... do we do something more fancy? Probably just stack them, but we'll
have to actually decide.