MidiRouter: Fix 5-pin duplicated device, adjust noteChanged signal for more verbosity
The first thing is not really super impactful, but it's nice to have
things working like they're supposed to. It does, however, also ensure
that we don't have two entries in the Recording popup, which just looks
a bit silly.

Also, adjusted the noteChanged signal, which now includes relevant
information about ports and what devices things arrived from and so on,
whenever, well, relevant.