PatternModel: Add next-step step entry property, treat vel=0 as trigless
Trigless in short means "don't play this note", which effectively is
what it means anyway (many synths treat velocity 0 as identical to a
note-off message), so we can use that for... fun things like steps with
nothing playing but which have fun logic on them, like...

The property next-step, which adds the ability to decide what step
should be played after the current one (in place of the logical next
step in the pattern). This will also obey the probability on that note,
which means it can be used not only to create non-even pattern lengths
and patterns with an attack section that only plays once, but also in a
probabilistic way, to make the playback jump around in fun ways.