ClipAudioSource: Add PlaybackStyle enum and matching properties
This is supposed to replace the data setters for looping and granular
mode, in favour of a single field which describes how playback should be
done for this sample. It also introduces the One-shot concept, and that
is kind of why the enum happened (rather than just yet another boolean
property). The actual playback logic is in other commits.