KeyScales: Add shorthand conversion functions for scale,pitch,octave
These functions ensure that we can store a stable version of the three
values to our patterns, and restore that data even if we decide to make
some changes later (such as, and notably, the order of our scales, the
number and complexity of which may well change over time). We don't
currently have any way of doing micro-tonal scales, but who knows, that
might well happen in the future. In the meantime, though, there are
other interesting scales we'll want to also support, so... a step
towards that is to be stable if we add new ones.