KeyScales: Add a bunch of new scales (and modes)
Some of these are duplicates, arguably (as in, they're the same interval
between pitches), but people might know that, but they'd also want to
say it's in, say, Flamenco, instead of Double Harmonic, as it's a
suggestion of intent more than technical lists of note values. So, more
fun.

It's a shame we can't do microtonal things without some huge upheaval,
but... yeah, this'll do for now.