KeyScales: Rework our lists a touch
Due to a misunderstanding about how QHash works, i was expecting things
to be in order, which of course they are not, given it's a QHash, which
id documented specifically to not keep its order, where QMap does. But,
rather than porting to that, instead we now use outright arrays, because
well... why not, really. We don't need the convenience of the others, so
we might as well. These are lookup tables, so let's treat them as such.