KeyScales: Adjust how transposeNote deals with off-scale notes
As the new documentation says, the first step of transposition will now
be to move the note onto the scale if it not already there. Otherwise,
the transposition works as it did before.

We also now have a noteIndices lookup table, to help us work out whether
a note is in fact on scale and at what index in the table, at O(1) speeds.
No more searching, just array access. Nice.