PatternModel: Make swing drummachine-style, and rework the live sequencer storage
Previously we were juggling a few different data structures explicitly,
and those have been all refactored into a struct named StepData, which
contains the step's midi buffers, its probability sequences, the
calculated swing offset, and an explicit field to say whether it is
valid or not, which makes a few things a touch simpler.

We also switched our swing logic from a 0 through 100 forward offset
style logic (where it would position forward only, rather than also
offering backward swing), to one which is akin to how drum machines will
perform the job (that is, any even step is positioned at the percentage
position between the previous and next steps that is the swing value,
meaning that the default is 50 (the centre-point), 1 is the lowest
(almost on top of the previous step) and 99 is the highest (almost on
top of the next step). We don't allow 0 and 100, because while we could
technically handle multiple same on notes (they would just be on
different channels), not all of our engines would support that, and it's
just kind of something we offer other solutions for (essentially just
consider that an "accent", so just turn up the velocity instead). So, in
short, new swing actually works as expected, and also, importantly, just
generally works :)