Handle note-on at vel-0 as a note-off message
There is, out there in the wide world of midi, a number of controllers
which instead of sending out note-off messages send out a note-on
message with velocity 0. This has resulted in there being a sort of
defacto assumption that note-on with velocity 0 is not at all a note-on
message that can then be reworked with aftertouch, but rather outright a
note-off message. A little awkward and weird, but it's a thing, so let's
just deal with that.

This also highlights a few discrepancies in comparisons and variable
naming. That will want sorting out at some point.