PatternModel: Remove old midiChannel logic in favour of sketchpadTrack
This is fairly invasive, but has wanted to happen for a fair while at
this point. We used to have a direct logical connection between a midi
channel and the sketchpad track that might be associated with, for
routing purposes, but since some time now, we've separated those out
entirely (in favour of a strict disconnection, to allow for any
arbitrary amount of midi channel events to happen on any sketchpad
track, to allow for mpe style data to be used). So, in one swoop, this
bunch of changes removes that old midiChannel logic, and changes the
incorrectly named channelIndex field to sketchpadTrack.

This was spurred on by the fact that unbouncing a sketch would load
incorrect midi channel information into a pattern, which would then
attempt to play the notes into the wrong sketchpad track (the one the
sketch had been created from), resulting in some very strange noises
happening.