PatternModel: Switch availableBars logic to a patternLength property
The patternLength property operates on a number of steps in a pattern,
and availableBars now is just the bar number of the last step in the
pattern. For example, a pattern with 16 steps will have 1 bar, and a
pattern with 17 steps will have 2 (because step 17 is the first step in
bar number 2).