PatternModel: Switch to an explicit step length setting
The old noteLength property was effectively a kind of symbolic way of
defining the length of a step. It came with us from very long ago when
we didn't have quite as much precision in our playback system, and had
both ended up too imprecise for our needs, and also just outright
confusing to use in general. The new system is an explicit setting of
how long each step is, by a number subdivision of a bar/beat (currently
equivalent to the sequencer's playback subdivision, but arguably that
might change at some point in the future, hence the double value in
favour of an integer).