DiskWriter: Move the lastRecordingFrame and isRecording logic in here
This really is where it belongs, and while that signal could now
technically be listened to from somewhere that causes a bunch of work to
happen in the playback loop... well, correctness is important, and very
much so here. We ourselves only listen to it through a queued
connection, so that part should be all safe and whatnot.