Add Compressor logic from IEM's plugin suite (GPL)
Institute of Electronic Music and Acoustics (IEM) has a very wide
ranging and capable set of code, one of which being this compressor
code. Including it here directly rather than adding the whole thing or
whatnot, but that's something to look at later perhaps. Lots of
niftiness in that repository.