JackPassthrough: Add equaliser/compressor logic
These two are as yet incomplete, but to allow for development to be done
on them, it's time to get this in. Have done a basic test of the
equaliser, and the default settings need some work, but tweaking is
easier when it's actually in.

It also exposed a fun thing where i think out peak decay is probably a
touch too long ;)