Move JackPassthrough visualisation logic from imageprovider to an item
While the item is still a QQuickPaintedItem, is is vastly better of a
solution than the image provider solution, which was both overly
intricate and invasive, as well as just very slow.

Kind of makes me think maybe we want to do something like this for our
patterns as well... the logic's kind of similar, could be nice.