ProcessWrapper: More trustworthy blocking call logic
Something seems to be breaking the observer pattern in QProcess, but...
we'll just have to deal with that. This works around the issue, and
while it means we'll only be able to get either stderr or stdout from a
blocking call, it still should be useful enough to be useful.