ProcessWrapper: Add a waitForOutput function
The function takes a regexp string and waits for that to happen, or
until an optional timeout occurs. The function returns either timeout or
success (enum values) depending on the outcome. A failure state also
exists, but is not currently used (may be used for e.g. a process crash
while waiting).