ProcessWrapper: Switch to using KPty{Device,Process}
Since things like jalv implements an interactive shell for, well,
interacting with it, and the implementation of that interaction in
zyngine expects that to be a thing, rather than forcing a new method
onto that code (which would be unwieldy to port), let's actually
implement a usable interactive mode for ProcessWrapper. KPty makes that
a thing that we can do. Nice.